Remove old pipeline
